Refactor core checklist
